Correct GUID structure.

+++ Loader/FileSystem/FileSystem.h
@@ -12,8 +12,8 @@ using LBARange = BasicRange<u64>;
 
 struct GUID {
     u32 data1;
-    u32 data2;
-    u32 data3;
+    u16 data2;
+    u16 data3;
     u8  data4[8];
 };
